Move `pkg/ruler/base/rulespb` to `pkg/ruler/rulespb`.
