+++ pkg/ruler/rulespb/compat.go
